requirements.txt
google-generativeai
nltk
requests
youtube-transcript-api
scikit-learn